+++ pxl-javax/src/main/java/io/github/hclimkr/pxl/internal/codec/PxlLocalDateTimeCodec.java
@@ -10,7 +10,6 @@ import io.github.hclimkr.pxl.internal.support.PxlDateCellSupport;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.CellType;
-import org.apache.poi.ss.usermodel.DateUtil;
 
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
@@ -58,13 +57,8 @@ final class PxlLocalDateTimeCodec {
         final CellType cellType = cell.getCellType();
         switch (cellType) {
             case NUMERIC:
-                final double numericValue = cell.getNumericCellValue();
                 try {
-                    if (DateUtil.isCellDateFormatted(cell)) {
-                        localDateTimeValue = cell.getLocalDateTimeCellValue();
-                    } else {
-                        localDateTimeValue = DateUtil.getLocalDateTime(numericValue);
-                    }
+                    localDateTimeValue = PxlDateCellSupport.readNumericCellAsLocalDateTime(cell, "LocalDateTime");
                 } catch (NumberFormatException numberFormatException) {
                     throw new PxlCellCodecException(PxlI18nDiagnostic.get(PxlI18nDiagnosticKeys.CODEC_PARSE_INVALID, String.valueOf(cell), "LocalDateTime"), numberFormatException);
                 }