+++ pxl-javax/src/main/java/io/github/hclimkr/pxl/internal/codec/PxlOffsetDateTimeCodec.java
@@ -10,7 +10,6 @@ import io.github.hclimkr.pxl.internal.support.PxlDateCellSupport;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.CellType;
-import org.apache.poi.ss.usermodel.DateUtil;
 
 import java.time.DateTimeException;
 import java.time.OffsetDateTime;
@@ -59,13 +58,8 @@ final class PxlOffsetDateTimeCodec {
         final CellType cellType = cell.getCellType();
         switch (cellType) {
             case NUMERIC:
-                final double numericValue = cell.getNumericCellValue();
                 try {
-                    if (DateUtil.isCellDateFormatted(cell)) {
-                        offsetDateTimeValue = cell.getLocalDateTimeCellValue().atZone(ZoneId.systemDefault()).toOffsetDateTime();
-                    } else {
-                        offsetDateTimeValue = DateUtil.getLocalDateTime(numericValue).atZone(ZoneId.systemDefault()).toOffsetDateTime();
-                    }
+                    offsetDateTimeValue = PxlDateCellSupport.readNumericCellAsLocalDateTime(cell, "OffsetDateTime").atZone(ZoneId.systemDefault()).toOffsetDateTime();
                 } catch (NumberFormatException numberFormatException) {
                     throw new PxlCellCodecException(PxlI18nDiagnostic.get(PxlI18nDiagnosticKeys.CODEC_PARSE_INVALID, String.valueOf(cell), "OffsetDateTime"), numberFormatException);
                 }