+++ pxl-javax/src/main/java/io/github/hclimkr/pxl/internal/codec/PxlZonedDateTimeCodec.java
@@ -10,7 +10,6 @@ import io.github.hclimkr.pxl.internal.support.PxlDateCellSupport;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.poi.ss.usermodel.Cell;
 import org.apache.poi.ss.usermodel.CellType;
-import org.apache.poi.ss.usermodel.DateUtil;
 
 import java.time.DateTimeException;
 import java.time.ZoneId;
@@ -60,13 +59,8 @@ final class PxlZonedDateTimeCodec {
         final CellType cellType = cell.getCellType();
         switch (cellType) {
             case NUMERIC:
-                final double numericValue = cell.getNumericCellValue();
                 try {
-                    if (DateUtil.isCellDateFormatted(cell)) {
-                        zonedDateTimeValue = cell.getLocalDateTimeCellValue().atZone(zoneId);
-                    } else {
-                        zonedDateTimeValue = DateUtil.getLocalDateTime(numericValue).atZone(zoneId);
-                    }
+                    zonedDateTimeValue = PxlDateCellSupport.readNumericCellAsLocalDateTime(cell, "ZonedDateTime").atZone(zoneId);
                 } catch (NumberFormatException numberFormatException) {
                     throw new PxlCellCodecException(PxlI18nDiagnostic.get(PxlI18nDiagnosticKeys.CODEC_PARSE_INVALID, String.valueOf(cell), "ZonedDateTime"), numberFormatException);
                 }